backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
reportlab==4.0.9
Pillow==10.2.0

# AI & Agent Framework
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.13
langgraph==0.0.20

# Utilities
python-dotenv==1.0.0
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
aiofiles==23.2.1
emails==0.6

# Evaluation & Monitoring
langsmith==0.0.77
datasets==2.16.1
